Fix -Wc++11-narrowing compile errors with llvm-mingw.
(cherry picked from commit 02b291363819b5da8c12e1b401b08a8917bcacc5)

+++ dom/plugins/ipc/PluginInstanceParent.cpp
@@ -1370,7 +1370,7 @@ PluginInstanceParent::UpdateScrollCapture(bool& aRequestNewCapture)
     container->SetCurrentImages(imageList);
 
     // Invalidate our area in the page so the image gets flushed.
-    NPRect nprect = {0, 0, targetSize.width, targetSize.height};
+    NPRect nprect = {0, 0, static_cast<uint16_t>(targetSize.width), static_cast<uint16_t>(targetSize.height)};
     RecvNPN_InvalidateRect(nprect);
 
     return true;
